pyproject.toml
[tool.poetry]
name = "nonebot-plugin-resolver"
version = "1.2.28"
description = "NoneBot2链接分享解析器插件。解析视频、图片链接/小程序插件，tiktok、bilibili、twitter等实时发送！"
authors = ["zhiyu1998 <renzhiyu0416@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot-plugin-resolver"}]
repository = "https://github.com/zhiyu1998/nonebot-plugin-resolver"
keywords = ["nonebot", "nonebot2", "resolver"]

[tool.poetry.dependencies]
python = ">=3.9"
aiohttp = "^3.7"
httpx = ">=0.23"
PyExecJS = "^1.5.1"
bilibili-api-python = ">=16.2.0"
aiofiles = ">=0.8.0"
yt-dlp = ">=2024.11.18"
nonebot2 = ">=2.0.0-beta.5"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot_plugin_localstore = ">=0.6.0"
nonebot-plugin-apscheduler = ">=0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"